requirements.txt
influxdb==5.3.1
fastapi==0.110.1
uvicorn==0.29.0
pydantic-core==2.18.1
pydantic-settings==2.2.1
python-multipart==0.0.9
mypy==1.9.0
types-psycopg2==2.9.21.20240417
psycopg2-binary==2.9.9